Create ConvertVideotoAudioWav
This is basic code for ffmpeg, which is capable of removing video, and the converting that resultant audio file to .wav. The one problem I can see is that we might need to somehow have ffmpeg preinstalled. If this ends up not working, I can try finding a different way of accomplishing this task.