Create GUIandConversion
This class is responsible for forming the basic setup of the GUI, along with a button that will access the user's video files. It will take these files and convert them into .wav format, then save them into the application's folder. All of the imports should be automatically on your computer except for ffmpeg.